requirements.txt
Flask==3.1.1
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
simple-websocket==1.1.0
SQLAlchemy==2.0.36
Werkzeug==3.1.3
